% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FpopPSD.R
\name{plot_seg}
\alias{plot_seg}
\title{plot_seg}
\usage{
plot_seg(cp, y, x)
}
\arguments{
\item{cp}{a vector of changepoints}

\item{y}{a vector of data}

\item{x}{a vector of data abscissa}
}
\description{
plot the changepoints and the mean of the segments on the data.
}
\examples{
library(ggplot2)
path_to_FpopPSD <- path.package("FpopPSD")
df <- read.table(paste0(path_to_FpopPSD,"/wellLogData.txt"), header = TRUE)
n <- length(df$y)
beta <- 2.5
alpha <- 10 + beta * log(n)
res <- FpopPSD::fpopPSD(df$y, beta, alpha)
FpopPSD::plot_seg(res$changepoints, df$y, df$time)
}
