% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FpopPSD.R
\name{plot_seg}
\alias{plot_seg}
\title{plot_seg}
\usage{
plot_seg(cp, y, x)
}
\arguments{
\item{cp}{a vector of changepoints}

\item{y}{a vector of data}

\item{x}{a vector of data abscissa}
}
\description{
plotting the changepoints end the mean of the segments on the data.
}
